feat: change track censoring approach
Instead of adding random noise to jitter the absolute training time and
distance values, simply adjust them so that the first data point for
each track is at 0 seconds / 0 meters.
This will simplify the censoring conceptually (though not code-wise),
at the cost of shortening all tracks by a certain time and distance
(whereas the Gaussian jitter should have left the average training
time and distance more or less intact).
This is a more drastic and blunt approach to censoring, but should
reduce the risk of accidentally leaking information.