feat: add presentation
Add the article for the presentation, and add required packages to the
package description file.
The presentation should give an overview of things attempted so far, to
get feedback and inputs from the rest of the class.
Together with the feedback it will be the basis for the final report.